// src/lib/covalent/types.ts
export interface ContractMetadata {
  contract_decimals: number;
  contract_name: string;
  contract_ticker_symbol: string;
  contract_address: string;
  supports_erc: string[] | null;
  logo_url: string;
}

export interface AddressWithHistoricalPricesItem extends ContractMetadata {
  update_at: string;
  quote_currency: string;
  prices: HistoricalPriceItem[];
  items: any;
}

export interface HistoricalPriceItem {
  contract_metadata: ContractMetadata;
  date: string;
  price: number;
}

export interface BalanceResponseType {
  address: string;
  updated_at: string;
  next_update_at: string;
  quote_currency: string;
  chain_id: number;
  items: WalletBalanceItem[];
  pagination?: AppliedPagination | null /* ? the address one has it, the address one in pools doesn't */;
}

export type INFTMetadata = any;

export type AssetType = 'cryptocurrency' | 'stablecoin' | 'nft' | 'dust';

export interface WalletBalanceItem extends ContractMetadata {
  last_transferred_at: string;
  type: AssetType;
  balance: string;
  balance_24h: string;
  quote_rate: number | null;
  quote_rate_24h: number | null;
  quote: number | null;
  quote_24h: number | null;
  nft_data: INFTMetadata[] | null;
}

export interface HistoricalPortfolioResponseItemTokenBalance {
  balance: string;
  quote: number | null;
}

export interface HistoricalPortfolioResponseItemToken {
  timestamp: string;
  quote_rate: number;
  open: HistoricalPortfolioResponseItemTokenBalance;
  high: HistoricalPortfolioResponseItemTokenBalance;
  low: HistoricalPortfolioResponseItemTokenBalance;
  close: HistoricalPortfolioResponseItemTokenBalance;
}

export interface HistoricalPortfolioResponseItem extends ContractMetadata {
  holdings: HistoricalPortfolioResponseItemToken[];
}

export interface HistoricalPortfolioResponse {
  address: string;
  updated_at: string;
  next_update_at: string;
  quote_currency: string;
  chain_id: number;
  items: HistoricalPortfolioResponseItem[];
  pagination: AppliedPagination | null;
}

export interface AppliedPagination {
  has_more: boolean;
  page_number: number;
  page_size: number;
  total_count: number | null;
}

export interface DecodedParamItem {
  name: string;
  type: string;
  indexed: boolean;
  decoded: boolean;
  value: string;
}

export interface TransactionResponse {
  address: string;
  updated_at: string;
  next_update_at: string;
  quote_currency: string;
  chain_id: number;
  items: BlockTransactionWithContractTransfers[];
  pagination: AppliedPagination | null;
}

export interface DecodedItem {
  name: string;
  signature: string;
  params: DecodedParamItem[];
}

export interface LogEventItem {
  block_signed_at: string;
  block_height: number;
  tx_offset: number;
  log_offset: number;
  tx_hash: string;
  raw_log_topics: any;
  sender_contract_decimals: number | null;
  sender_name: string | null;
  sender_contract_ticker_symbol: string | null;
  sender_address: string;
  sender_address_label: string | null;
  sender_logo_url: string | null;
  raw_log_data: string;
  decoded: DecodedItem;
}

export interface BlockTransactionWithLogEvents {
  block_signed_at: string;
  block_height: number;
  tx_hash: string;
  tx_offset: number;
  successful: boolean;
  from_address: string;
  from_address_label: string | null;
  to_address: string;
  to_address_label: string | null;
  value: string;
  value_quote: number;
  gas_offered: number;
  gas_spent: number;
  gas_price: number;
  fees_paid: string | null;
  gas_quote: number;
  gas_quote_rate: number;
  log_events: LogEventItem[];
}

export interface SingleTransactionResponse {
  updated_at: string;
  items: BlockTransactionWithLogEvents[];
  pagination: AppliedPagination | null;
}

export interface MethodCallsForTransfers {
  sender_address: string;
  method: string;
}

export interface TransferResponse {
  address: string;
  updated_at: string;
  next_update_at: string;
  quote_currency: string;
  items: BlockTransactionWithContractTransfers[];
  pagination: AppliedPagination | null;
}

export type TokenTransferItemType = 'IN' | 'OUT';

export interface TokenTransferItem {
  block_signed_at: string;
  tx_hash: string;
  from_address: string;
  from_address_label: string | null;
  to_address: string;
  to_address_label: string | null;
  contract_decimals: number;
  contract_name: string;
  contract_ticker_symbol: string;
  contract_address: string;
  logo_url: string;
  transfer_type: TokenTransferItemType;
  delta: string;
  balance: string | null;
  quote_rate: number;
  delta_quote: number;
  balance_quote: number | null;
  method_calls: MethodCallsForTransfers[] | null;
}

export interface BlockTransactionWithContractTransfers {
  block_signed_at: string;
  block_height: number;
  tx_hash: string;
  tx_offset: number;
  successful: boolean;
  from_address: string;
  from_address_label: string | null;
  to_address: string;
  to_address_label: string | null;
  value: string;
  value_quote: number;
  gas_offered: number;
  gas_spent: number;
  gas_price: number;
  fees_paid: number | null;
  gas_quote: number;
  gas_quote_rate: number;
  transfers: TokenTransferItem;
}

export interface Block {
  signed_at: string;
  height: number;
}

export interface SingleBlockResponse {
  updated_at: string;
  items: Block[];
  pagination: AppliedPagination | null;
}
// ...
export interface EventsListResponseType {
  updated_at: string;
  items: LogEventItem[];
  pagination: AppliedPagination | null;
}
// ...
export interface TokenHolder extends ContractMetadata {
  address: string;
  balance: string;
  total_supply: string;
  block_height: number;
}

export interface TokenHolderResponse {
  updated_at: string;
  items: TokenHolder[];
  pagination: AppliedPagination | null;
}

export interface TokenHolderDiff {
  token_holder: string;
  prev_balance: string;
  prev_block_height: number;
  next_balance: string;
  next_block_height: number;
  diff: string;
}

export interface TokenHolderDiffResponse {
  updated_at: string;
  items: TokenHolderDiff[];
  pagination: AppliedPagination | null;
}
// ...
export interface TokenIdResponse extends ContractMetadata {
  token_id: string;
}
// ...
export interface NftTransactionsResponse {
  contract_decimals: number;
  contract_name: string;
  contract_ticker_symbol: string;
  contract_address: string;
  supports_erc: string[];
  logo_url: string;
  type: AssetType;
  nft_transactions: BlockTransactionWithLogEvents[];
}

export interface NftTransactionsResponseType {
  updated_at: string;
  items: NftTransactionsResponse[];
  pagination: AppliedPagination | null;
}

export interface NftMetadataResponse extends ContractMetadata {
  type: AssetType;
  nft_data: INFTMetadata;
}

export interface NftMetadataResponseType {
  updated_at: string;
  items: NftMetadataResponse[];
  pagination: AppliedPagination | null;
}
// ...
export interface GenericChainInfoDisplay {
  name: string;
  chain_id: string;
  is_testnet: boolean;
  db_schema_name: string;
  label: string;
  logo_url: string;
}

export interface AllChainInfoResponse {
  updated_at: string;
  items: GenericChainInfoDisplay;
  pagination: AppliedPagination | null;
}

export interface ChainStatusResponse {
  updated_at: string;
  items: GenericChainInfoStatusDisplay[];
  pagination: AppliedPagination | null;
}

export interface GenericChainInfoStatusDisplay {
  name: string;
  chain_id: string;
  is_testnet: boolean;
  logo_url: string;
  synced_block_height: number;
  synced_blocked_signed_at: string;
}